init: added validators and some fixes

+++ Library_DAL/Entities/Models/Book.cs
@@ -11,7 +11,7 @@ namespace Libriary_DAL.Entities.Models
         public int BookId {  get; set; }
         public int GenreId { get; set; }
         public int AuthorId { get; set; }
-        public int ISBN { get; set; }
+        public long ISBN { get; set; }
         public string? Title { get; set; }
         public string? Description { get; set; }
